Revert "refactor bigquery execution (#865)"
This reverts commit 84c1b99b1cc193abd337adffc12bc5155c133a0c.
