--- legend-engine-xt-relationalStore-bigquery-execution/src/main/java/org/finos/legend/engine/plan/execution/stores/relational/BigQueryConnectionExtension.java
@@ -1,105 +1,0 @@
-//  Copyright 2022 Goldman Sachs
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-
-package org.finos.legend.engine.plan.execution.stores.relational;
-
-import org.eclipse.collections.api.block.function.Function2;
-import org.eclipse.collections.api.factory.Lists;
-import org.eclipse.collections.api.list.MutableList;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.ConnectionKey;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.authentication.AuthenticationStrategy;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.authentication.strategy.OAuthProfile;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.authentication.strategy.keys.AuthenticationStrategyKey;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.DatabaseManager;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.commands.RelationalDatabaseCommands;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.bigquery.BigQueryCommands;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.bigquery.BigQueryManager;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.ds.DataSourceSpecification;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.ds.DataSourceSpecificationKey;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.ds.specifications.BigQueryDataSourceSpecification;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.ds.specifications.keys.BigQueryDataSourceSpecificationKey;
-import org.finos.legend.engine.plan.execution.stores.relational.connection.manager.strategic.StrategicConnectionExtension;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.RelationalDatabaseConnection;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.authentication.AuthenticationStrategyVisitor;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.specification.BigQueryDatasourceSpecification;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.specification.DatasourceSpecificationVisitor;
-
-import java.util.List;
-import java.util.function.Function;
-
-public class BigQueryConnectionExtension implements RelationalConnectionExtension, StrategicConnectionExtension
-{
-    @Override
-    public MutableList<DatabaseManager> getAdditionalDatabaseManager()
-    {
-        return Lists.mutable.of(new BigQueryManager());
-    }
-
-    @Override
-    public Boolean visit(StreamResultToTempTableVisitor visitor, RelationalDatabaseCommands databaseCommands)
-    {
-        if (databaseCommands instanceof BigQueryCommands)
-        {
-            throw new UnsupportedOperationException("not yet implemented");
-        }
-        return null;
-    }
-
-    @Override
-    public AuthenticationStrategyVisitor<AuthenticationStrategyKey> getExtraAuthenticationKeyGenerators()
-    {
-        return authenticationStrategy -> null;
-    }
-
-    @Override
-    public AuthenticationStrategyVisitor<AuthenticationStrategy> getExtraAuthenticationStrategyTransformGenerators(List<OAuthProfile> oauthProfiles)
-    {
-        return authenticationStrategy -> null;
-    }
-
-    @Override
-    public Function<RelationalDatabaseConnection, DatasourceSpecificationVisitor<DataSourceSpecificationKey>> getExtraDataSourceSpecificationKeyGenerators(int testDbPort)
-    {
-        return connection -> datasourceSpecification ->
-        {
-            if (datasourceSpecification instanceof BigQueryDatasourceSpecification)
-            {
-                BigQueryDatasourceSpecification bigQueryDatasourceSpecification = (BigQueryDatasourceSpecification) datasourceSpecification;
-                return new BigQueryDataSourceSpecificationKey(
-                        bigQueryDatasourceSpecification.projectId,
-                        bigQueryDatasourceSpecification.defaultDataset,
-                        bigQueryDatasourceSpecification.proxyHost,
-                        bigQueryDatasourceSpecification.proxyPort);
-            }
-            return null;
-        };
-    }
-
-    @Override
-    public Function2<RelationalDatabaseConnection, ConnectionKey, DatasourceSpecificationVisitor<DataSourceSpecification>> getExtraDataSourceSpecificationTransformerGenerators(Function<RelationalDatabaseConnection, AuthenticationStrategy> authenticationStrategyProvider)
-    {
-        return (connection, connectionKey) -> datasourceSpecification ->
-        {
-            if (datasourceSpecification instanceof BigQueryDatasourceSpecification)
-            {
-                return new BigQueryDataSourceSpecification(
-                        (BigQueryDataSourceSpecificationKey) connectionKey.getDataSourceSpecificationKey(),
-                        new BigQueryManager(),
-                        authenticationStrategyProvider.apply(connection)
-                );
-            }
-            return null;
-        };
-    }
-}
