--- legend-engine-xt-relationalStore-bigquery-execution/src/main/resources/META-INF/services/org.finos.legend.engine.plan.execution.stores.relational.connection.ConnectionExtension
@@ -1,1 +1,0 @@
-org.finos.legend.engine.plan.execution.stores.relational.BigQueryConnectionExtension
