--- legend-engine-xt-relationalStore-bigquery-grammar/src/main/antlr4/org/finos/legend/engine/language/pure/grammar/from/antlr4/connection/BigQueryLexerGrammar.g4
@@ -1,9 +1,0 @@
-lexer grammar BigQueryLexerGrammar;
-
-import CoreLexerGrammar;
-
-BIGQUERY_DSP:                               'BigQuery';
-PROJECT:                                    'projectId';
-DATASET:                                    'defaultDataset';
-PROXYHOST:                                  'proxyHost';
-PROXYPORT:                                  'proxyPort';
