--- legend-engine-xt-relationalStore-bigquery-grammar/src/main/resources/META-INF/services/org.finos.legend.engine.language.pure.grammar.from.IRelationalGrammarParserExtension
@@ -1,1 +1,0 @@
-org.finos.legend.engine.language.pure.grammar.from.BigQueryGrammarParserExtension
