--- legend-engine-xt-relationalStore-bigquery-grammar/src/main/resources/META-INF/services/org.finos.legend.engine.language.pure.grammar.to.extension.PureGrammarComposerExtension
@@ -1,1 +1,0 @@
-org.finos.legend.engine.language.pure.grammar.to.BigQueryGrammarComposerExtension
