--- legend-engine-xt-relationalStore-bigquery-protocol/src/main/java/org/finos/legend/engine/protocol/pure/v1/BigQueryProtocolExtension.java
@@ -1,38 +1,0 @@
-// Copyright 2021 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.finos.legend.engine.protocol.pure.v1;
-
-import org.eclipse.collections.api.block.function.Function0;
-import org.eclipse.collections.impl.factory.Lists;
-import org.finos.legend.engine.protocol.pure.v1.extension.ProtocolSubTypeInfo;
-import org.finos.legend.engine.protocol.pure.v1.extension.PureProtocolExtension;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.specification.BigQueryDatasourceSpecification;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.specification.DatasourceSpecification;
-
-import java.util.List;
-
-public class BigQueryProtocolExtension implements PureProtocolExtension
-{
-    @Override
-    public List<Function0<List<ProtocolSubTypeInfo<?>>>> getExtraProtocolSubTypeInfoCollectors()
-    {
-        return Lists.fixedSize.with(() -> Lists.fixedSize.with(
-                //DatasourceSpecification
-                ProtocolSubTypeInfo.newBuilder(DatasourceSpecification.class)
-                        .withSubtype(BigQueryDatasourceSpecification.class, "bigQuery")
-                        .build()
-        ));
-    }
-}
