--- legend-engine-xt-relationalStore-bigquery-protocol/src/main/resources/META-INF/services/org.finos.legend.engine.protocol.pure.v1.extension.PureProtocolExtension
@@ -1,1 +1,0 @@
-org.finos.legend.engine.protocol.pure.v1.BigQueryProtocolExtension
