+++ legend-engine-xt-relationalStore-executionPlan-connection-authentication-default/src/main/java/org/finos/legend/engine/authentication/LegendDefaultDatabaseAuthenticationFlowProvider.java
@@ -43,7 +43,7 @@ public final class LegendDefaultDatabaseAuthenticationFlowProvider extends Abstr
     {
         return Lists.immutable.of(
                 new BigQueryWithGCPApplicationDefaultCredentialsFlow(),
-                new BigQueryWithGCPWorkloadIdentityFederationFlow(databaseAuthenticationFlowProviderConfiguration.getAwsConfig(), databaseAuthenticationFlowProviderConfiguration.getGcpWorkloadConfig()),
+                new BigQueryWithGCPWorkloadIdentityFederationFlow(databaseAuthenticationFlowProviderConfiguration),
                 new DatabricksWithApiTokenFlow(),
                 new H2StaticWithTestUserPasswordFlow(),
                 new SnowflakeWithKeyPairFlow(),
