--- legend-engine-xt-relationalStore-executionPlan-connection-authentication/src/main/java/org/finos/legend/engine/authentication/cloud/GCPWorkloadConfig.java
@@ -1,49 +1,0 @@
-// Copyright 2021 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.finos.legend.engine.authentication.cloud;
-
-public class GCPWorkloadConfig
-{
-    public String getProjectNumber()
-    {
-        return projectNumber;
-    }
-
-    public String getPoolId()
-    {
-        return poolId;
-    }
-
-    public String getProviderId()
-    {
-        return providerId;
-    }
-
-    public GCPWorkloadConfig(String projectNumber, String poolId, String providerId)
-    {
-        this.projectNumber = projectNumber;
-        this.poolId = poolId;
-        this.providerId = providerId;
-    }
-
-    private String projectNumber;
-    private String poolId;
-    private String providerId;
-
-    public GCPWorkloadConfig()
-    {
-        // jackson
-    }
-}
