+++ legend-engine-xt-relationalStore-executionPlan-connection/src/main/resources/META-INF/services/org.finos.legend.engine.plan.execution.stores.relational.connection.ConnectionExtension
@@ -1,0 +1,1 @@
+org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.bigquery.BigQueryConnectionExtension
