+++ legend-engine-xt-relationalStore-executionPlan/src/main/java/org/finos/legend/engine/plan/execution/stores/relational/StreamResultToTempTableVisitor.java
@@ -29,6 +29,7 @@ import org.finos.legend.engine.plan.execution.stores.relational.connection.drive
 import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.commands.IngestionMethod;
 import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.commands.RelationalDatabaseCommands;
 import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.commands.RelationalDatabaseCommandsVisitor;
+import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.bigquery.BigQueryCommands;
 import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.databricks.DatabricksCommands;
 import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.h2.H2Commands;
 import org.finos.legend.engine.plan.execution.stores.relational.connection.driver.vendors.postgres.PostgresCommands;
@@ -90,6 +91,10 @@ public class StreamResultToTempTableVisitor implements RelationalDatabaseCommand
         if (databaseCommands instanceof H2Commands)
         {
             return visitH2((H2Commands) databaseCommands);
+        }
+        if (databaseCommands instanceof BigQueryCommands)
+        {
+            return visitBigQuery((BigQueryCommands) databaseCommands);
         }
         if (databaseCommands instanceof RedshiftCommands)
         {
@@ -206,6 +211,11 @@ public class StreamResultToTempTableVisitor implements RelationalDatabaseCommand
             streamResultToNewTarget(((RelationalResult) result).resultSet, connection, tableName, 100);
         }
         return true;
+    }
+
+    private Boolean visitBigQuery(BigQueryCommands bigQueryCommands)
+    {
+        throw new UnsupportedOperationException("not yet implemented");
     }
 
     private Boolean visitRedshift(RedshiftCommands redshiftCommands)
