+++ legend-engine-xt-relationalStore-grammar/src/main/antlr4/org/finos/legend/engine/language/pure/grammar/from/antlr4/connection/datasource/DataSourceSpecificationLexerGrammar.g4
@@ -29,6 +29,10 @@ ACCOUNTTYPE:                                'accountType';
 ORGANIZATION:                               'organization';
 ROLE:                                       'role';
 
+BIGQUERY_DSP:                               'BigQuery';
+PROJECT:                                    'projectId';
+DATASET:                                    'defaultDataset';
+
 DATABRICKS:                                 'Databricks';
 PROTOCOL:                                   'protocol';
 HTTP_PATH:                                  'httpPath';
