+++ legend-engine-xt-relationalStore-grammar/src/main/antlr4/org/finos/legend/engine/language/pure/grammar/from/antlr4/connection/datasource/DataSourceSpecificationParserGrammar.g4
@@ -113,6 +113,22 @@ snowflakeQuotedIdentifiersIgnoreCase:       QUOTED_IDENTIFIERS_IGNORE_CASE COLON
 ;
 dbRole:                                     ROLE COLON STRING SEMI_COLON
 ;
+
+bigQueryDatasourceSpecification:            BIGQUERY_DSP
+                                                BRACE_OPEN
+                                                    (
+                                                        projectId
+                                                        | defaultDataset
+                                                        | dbProxyHost
+                                                        | dbProxyPort
+                                                    )*
+                                                BRACE_CLOSE
+;
+projectId:                                    PROJECT COLON STRING SEMI_COLON
+;
+defaultDataset:                                    DATASET COLON STRING SEMI_COLON
+;
+
 region:                                     REGION COLON STRING SEMI_COLON
 ;
 endpointURL:                                 ENDPOINT_URL COLON STRING SEMI_COLON
