+++ pom.xml
@@ -92,10 +92,6 @@
         <module>legend-engine-xt-relationalStore-sqlserver-execution</module>
         <module>legend-engine-xt-relationalStore-sqlserver-pure</module>
 
-        <module>legend-engine-xt-relationalStore-bigquery-execution</module>
-        <module>legend-engine-xt-relationalStore-bigquery-protocol</module>
-        <module>legend-engine-xt-relationalStore-bigquery-grammar</module>
-
         <module>legend-engine-xt-avro</module>
         <module>legend-engine-xt-avro-pure</module>
 
@@ -691,21 +687,6 @@
             <dependency>
                 <groupId>org.finos.legend.engine</groupId>
                 <artifactId>legend-engine-xt-relationalStore-sqlserver-pure</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.finos.legend.engine</groupId>
-                <artifactId>legend-engine-xt-relationalStore-bigquery-execution</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.finos.legend.engine</groupId>
-                <artifactId>legend-engine-xt-relationalStore-bigquery-protocol</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.finos.legend.engine</groupId>
-                <artifactId>legend-engine-xt-relationalStore-bigquery-grammar</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
